Add files via upload
Create a function named unlockDoor that receives 3 parameters:
doorCode (number) - The code to unlock the door
numAttempts (number) - The number of attempts to guess the code
clue (string) - A clue to guess the code
The function will simulate an attempt to unlock a mysterious door by guessing the door code:
Use a for loop that iterates from 1 to numAttempts
In each iteration, check if the current iteration number equals the doorCode. If so, return "Door unlocked! The code was " + doorCode
If the loop finishes without guessing the code, return clue